Add more unit tests to the clean function
