+++ packages/block-editor/src/hooks/test/color.js
@@ -12,16 +12,8 @@ import { registerBlockType, unregisterBlockType } from '@wordpress/blocks';
  * Internal dependencies
  */
 import BlockEditorProvider from '../../components/provider';
-import { cleanEmptyObject } from '../utils';
 import { withColorPaletteStyles } from '../color';
 
-describe( 'cleanEmptyObject', () => {
-	it( 'should remove nested keys', () => {
-		expect( cleanEmptyObject( { color: { text: undefined } } ) ).toEqual(
-			undefined
-		);
-	} );
-} );
 
 describe( 'withColorPaletteStyles', () => {
 	const settings = {
